Removed unused class in About page along with the non-existent CarouselEvent function call.
Moved the background css items to the site.css file. The URL is kept in Index because we only want the Index page to show it and it only works if it is put in this file due to cshtml and path reasons.

Removed unnecessary comments and unused classes from site.css. Modified the @media rule to properly disply items on smaller screens (mobile devices).

Removed the old PirateKingBMainImage.webp and updated the 'About Carousel.txt' file as needed.